backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
transformers
tensorflow
tf-keras
